Added a test to test the dh keysize logic
