+++ dev/com.ibm.ws.ssl/test/com/ibm/ws/ssl/provider/SSLProviderTest.java
@@ -1,0 +1,66 @@
+/*******************************************************************************
+ * Copyright (c) 2011 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.ssl.provider;
+
+import static org.junit.Assert.assertEquals;
+
+import java.lang.reflect.Field;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+import test.common.SharedOutputManager;
+
+/**
+ * Test for SecurityDefaults class
+ */
+public class SSLProviderTest {
+
+    private static SharedOutputManager outputMgr;
+
+    @BeforeClass
+    public static void setUpBeforeClass() throws Exception {
+        outputMgr = SharedOutputManager.getInstance();
+        outputMgr.captureStreams();
+    }
+
+    @Before
+    public void setUp() throws Exception {
+        // Clear the system property before each test
+        System.clearProperty("jdk.tls.ephemeralDHKeySize");
+        
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        outputMgr.resetStreams();
+    }
+
+    @AfterClass
+    public static void tearDownAfterClass() throws Exception {
+        outputMgr.restoreStreams();
+    }
+
+    @Test
+    public void defaultDHKeySizeSetWhenNotSet() {
+        SecurityDefaults.ensureDhKeySize();
+        assertEquals("2048",
+            System.getProperty("jdk.tls.ephemeralDHKeySize"));
+    }
+
+}
+
